Naprawa ID, GET test dla reservations, check deleting reservations, and more ...

+++ src/main/java/app/managers/ProductManager.java
@@ -5,7 +5,6 @@ import app.repositories.ProductRepository;
 import jakarta.inject.Inject;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.function.Predicate;
@@ -15,19 +14,23 @@ public class ProductManager {
 	@Inject
 	private ProductRepository productRepository;
 
-	public Product create(int price) {
+	public Product create (int price) {
 		return productRepository.insert(new Product(price));
 	}
 
-	public void delete(int id) {
-		productRepository.delete(id);
+	public void delete (int id) throws Exception {
+		Product p = productRepository.get(id);
+		if (p.getCurrentReservations().isEmpty())
+			productRepository.delete(id);
+		else
+			throw new Exception("cannot delete product with ongoing reservations");
 	}
 
-	public Product modify(int id, Function<Product, Product> func) throws Exception {
+	public Product modify (int id, Function<Product, Product> func) throws Exception {
 		return productRepository.modify(id, func);
 	}
 
-	public Product get(int id) {
+	public Product get (int id) {
 		return productRepository.get(id);
 	}
 
@@ -40,11 +43,11 @@ public class ProductManager {
 		return null;
 	}
 
-	public HashMap<Integer, Product> getMap() {
+	public HashMap<Integer, Product> getMap () {
 		return productRepository.getMap();
 	}
 
-	public int getLength() {
+	public int getLength () {
 		return productRepository.getLenght();
 	}
